# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ItemA:
    properties:
      id:
        type: string
    type: object
  domain.ItemB:
    properties:
      id:
        type: string
    type: object
  http.ResponseError:
    properties:
      error: {}
      statusCode:
        type: integer
    type: object
host: localhost:8085
info:
  contact: {}
  description: This is a sample server.
  title: Service A Swagger Example API
  version: "1.0"
paths:
  /a-items:
    get:
      consumes:
      - application/json
      description: Return all stored items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ItemA'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Show all items
      tags:
      - itemA
    post:
      consumes:
      - application/json
      description: creates an item with given data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemA'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Creates an item
      tags:
      - itemA
  /a-items/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an item with given ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemA'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Deletes an item
      tags:
      - itemA
    get:
      consumes:
      - application/json
      description: get item by ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemA'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Show an item
      tags:
      - itemA
    put:
      consumes:
      - application/json
      description: Updates an item with given ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Updates an item
      tags:
      - itemA
  /b-items:
    get:
      consumes:
      - application/json
      description: Return all stored items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ItemB'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Show all items
      tags:
      - itemB
    post:
      consumes:
      - application/json
      description: creates an item with given data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemB'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Creates an item
      tags:
      - itemB
  /b-items/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an item with given ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemB'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Deletes an item
      tags:
      - itemB
    get:
      consumes:
      - application/json
      description: get item by ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ItemB'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Show an item
      tags:
      - itemB
    put:
      consumes:
      - application/json
      description: Updates an item with given ID
      parameters:
      - description: Item ID
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ResponseError'
      summary: Updates an item
      tags:
      - itemB
swagger: "2.0"